YOLO/requirements.txt
tensorflow==1.12.0
scipy==1.1.0
numpy==1.15.4
matplotlib==3.0.1
Keras==2.2.4
Pillow==5.3.0
